fix(error): move common function to "helpers/error/combinePathErrors"
