+++ lib/error/setOptionError.js
@@ -5,8 +5,8 @@
 'use strict';
 
 const MongooseError = require('./mongooseError');
-// const getConstructorName = require('../helpers/getConstructorName');
 const util = require('util');
+const combinePathErrors = require('../helpers/error/combinePathErrors');
 
 class SetOptionError extends MongooseError {
   /**
@@ -25,7 +25,7 @@ class SetOptionError extends MongooseError {
    * Console.log helper
    */
   toString() {
-    return _generateMessage(this);
+    return combinePathErrors(this);
   }
 
   /**
@@ -53,7 +53,7 @@ class SetOptionError extends MongooseError {
     }
 
     this.errors[key] = error;
-    this.message = _generateMessage(this);
+    this.message = combinePathErrors(this);
   }
 }
 
@@ -81,27 +81,6 @@ Object.defineProperty(SetOptionError.prototype, 'toJSON', {
 Object.defineProperty(SetOptionError.prototype, 'name', {
   value: 'SetOptionError'
 });
-
-/*!
- * ignore
- */
-
-function _generateMessage(err) {
-  const keys = Object.keys(err.errors || {});
-  const len = keys.length;
-  const msgs = [];
-  let key;
-
-  for (let i = 0; i < len; ++i) {
-    key = keys[i];
-    if (err === err.errors[key]) {
-      continue;
-    }
-    msgs.push(key + ': ' + err.errors[key].message);
-  }
-
-  return msgs.join(', ');
-}
 
 class SetOptionInnerError extends MongooseError {
   /**
