+++ lib/error/validation.js
@@ -7,6 +7,7 @@
 const MongooseError = require('./mongooseError');
 const getConstructorName = require('../helpers/getConstructorName');
 const util = require('util');
+const combinePathErrors = require('../helpers/error/combinePathErrors');
 
 class ValidationError extends MongooseError {
   /**
@@ -38,7 +39,7 @@ class ValidationError extends MongooseError {
    * Console.log helper
    */
   toString() {
-    return this.name + ': ' + _generateMessage(this);
+    return this.name + ': ' + combinePathErrors(this);
   }
 
   /**
@@ -66,7 +67,7 @@ class ValidationError extends MongooseError {
     }
 
     this.errors[path] = error;
-    this.message = this._message + ': ' + _generateMessage(this);
+    this.message = this._message + ': ' + combinePathErrors(this);
   }
 }
 
@@ -94,27 +95,6 @@ Object.defineProperty(ValidationError.prototype, 'toJSON', {
 Object.defineProperty(ValidationError.prototype, 'name', {
   value: 'ValidationError'
 });
-
-/*!
- * ignore
- */
-
-function _generateMessage(err) {
-  const keys = Object.keys(err.errors || {});
-  const len = keys.length;
-  const msgs = [];
-  let key;
-
-  for (let i = 0; i < len; ++i) {
-    key = keys[i];
-    if (err === err.errors[key]) {
-      continue;
-    }
-    msgs.push(key + ': ' + err.errors[key].message);
-  }
-
-  return msgs.join(', ');
-}
 
 /*!
  * Module exports
